Move filter tests
